feat: complete PeopleTable and PersonLink
