+++ src/App.tsx
@@ -2,7 +2,7 @@ import './App.scss';
 import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
 import { Home } from './pages/Home/Home';
 import { NotFoundPage } from './pages/PageNotFound/NotFoundPage';
-import { PeoplePage } from './pages/PeoplePage/PeoplePage';
+import { PeopleTable } from './pages/PeopleTable/PeopleTable';
 import { MainLayout } from './layout/MainLayout';
 import { useEffect, useState } from 'react';
 import { Person } from './types';
@@ -43,14 +43,14 @@ export const App = () => {
         <Route
           path="people"
           element={
-            <PeoplePage
+            <PeopleTable
               people={people}
               isLoading={isLoading}
               hasError={hasError}
             />
           }
         >
-          <Route path=":slug" />
+          <Route path=":slug" element={null} />
         </Route>
         <Route path="home" element={<Navigate to="/" replace />} />
         <Route path="*" element={<NotFoundPage />} />
